Cleanup the subscription handling
